Created pages folder and moved all pages to 'pages' folder
