+++ src/App.js
@@ -1,9 +1,9 @@
 import React, {Component } from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import Home from './components/Home';
-import { Volunteer } from './components/Volunteer';
-import { Login } from './components/Login';
-import { NoMatch } from './components/NoMatch';
+import Home from './pages/Home';
+import { Volunteer } from './pages/Volunteer';
+import { Login } from './pages/Login';
+import { NoMatch } from './pages/NoMatch';
 
 import Navbar from './components/NavTab';
 // import './App.css';
