--- src/components/Home.js
@@ -1,19 +1,0 @@
-import React from 'react';
-
-
-
-function Home() {
-  return (
-  <div>
-    <h2>Home Page</h2>
-    <p>This is a test to see if this will display on the page</p>
-    <p>This is a test to see if this will display on the page</p>
-    <p>This is a test to see if this will display on the page</p>
-
-  </div>
-  );
-  }
-
-
-  export default Home;
-
